+++ src/App.tsx
@@ -6,36 +6,22 @@ import Instructions from "./Instructions";
 import useLocalStorage from "./useLocalStorage";
 
 import "./App.scss";
-import { github, abbreviatedSha, authorDate, commitMessage } from "~build/info";
-import time from "~build/time";
+import Github from "./Github";
 
 function App() {
 	const headerScrollElement = useRef(null);
 	const diceScrollElement = useRef(null);
 	const scoreboardScrollElement = useRef(null);
-	const instructionsScrollElement = useRef(null);
 
-	const [hasVisited, setHasVisited] = useLocalStorage("hasVisited", "true");
+	const [hasVisited, setHasVisited] = useLocalStorage("hasVisited", "false");
 
 	useEffect(() => {
-		if (localStorage.getItem("hasVisited")) {
-			console.log("returning");
-		} else {
-			console.log("setting local storage");
-			localStorage.setItem("hasVisited", "true");
+		if (localStorage.getItem("hasVisited") == "false") {
+			setTimeout(() => localStorage.setItem("hasVisited", "true"), 10000);
 		}
 	}, []);
 
 	//const [theme, setTheme] = useLocalStorage("theme", "dark");
-
-	var localeOptions: object = {
-		year: "numeric",
-		month: "2-digit",
-		day: "numeric",
-		hour: "2-digit",
-		minute: "2-digit",
-		timeZoneName: "short",
-	};
 
 	const setScrollPosition = (ref: { current: { offsetTop: any } }) => {
 		window.scrollTo({
@@ -43,12 +29,6 @@ function App() {
 			behavior: "smooth",
 		});
 	};
-
-	const openInNewTab = (url: string | undefined) => {
-		window.open(url, "_blank", "noreferrer");
-	};
-
-	//setTimeout(() => setHasVisited("returningUser"), 5000);
 
 	return (
 		<div className="App">
@@ -67,20 +47,7 @@ function App() {
 				headerScrollElement={headerScrollElement}
 				hasVisited={hasVisited}
 			/>
-			<div
-				className="github-container"
-				onClick={() => openInNewTab(github ? github : undefined)}
-			>
-				<span className="github-link">github</span>
-			</div>
-			<div className="commit-container">
-				<span className="commit-message">
-					<div className="commit-message">
-						Latest update: {time.toLocaleString("en-US", localeOptions)}
-					</div>
-					{"[" + abbreviatedSha + "] " + commitMessage}
-				</span>
-			</div>
+			<Github />
 		</div>
 	);
 }
