+++ src/Instructions.scss
@@ -1,4 +1,24 @@
-.instructions {
+.instructions-desktop {
+	font-family: "Roboto", sans-serif;
+	font-size: x-large;
+	text-align: left;
+	position: absolute;
+	top: 15vh;
+	display: flex;
+	cursor: pointer;
+	background: white;
+	color: black;
+	border-color: #e1a6ff;
+	border-radius: 20px;
+	margin: auto;
+	height: 750px;
+	width: 750px;
+	overflow-y: auto;
+	border-width: 12px;
+	animation: fade-in 1s;
+}
+
+.instructions-mobile {
 	font-family: "Roboto", sans-serif;
 	font-size: x-large;
 	text-align: left;
@@ -6,7 +26,6 @@
 	top: 25vh;
 	display: flex;
 	cursor: pointer;
-	//z-index: 50;
 	background: white;
 	color: black;
 	border-color: #e1a6ff;
@@ -15,9 +34,22 @@
 	height: 300px;
 	width: 300px;
 	overflow-y: auto;
-	//width: 80vw;
-	border-width: 8px;
+	border-width: 12px;
 	animation: fade-in 1s;
+	padding-top: 10px;
+}
+
+::-webkit-scrollbar {
+	width: 20px;
+	height: 20px;
+	margin-top: 20px;
+}
+
+::-webkit-scrollbar-thumb {
+	border: 4px solid rgba(0, 0, 0, 0);
+	background-clip: padding-box;
+	border-radius: 15px;
+	background-color: #aaaaaa;
 }
 
 .help-button {
@@ -31,9 +63,6 @@
 	border-width: 8px;
 	border-radius: 20px;
 	margin-top: 50px !important;
-	// button:hover {
-	// 	border-color: aquamarine !important;
-	// }
 }
 
 button:focus,
